// apps/frontend/src/gen/vibesync/common/v1/common_pb.ts
export class Id extends Message<Id> {
  /**
   * @generated from field: string value = 1;
   */
  value = "";

  constructor(data?: PartialMessage<Id>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "vibesync.common.v1.Id";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "value", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Id {
    return new Id().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Id {
    return new Id().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Id {
    return new Id().fromJsonString(jsonString, options);
  }

  static equals(a: Id | PlainMessage<Id> | undefined, b: Id | PlainMessage<Id> | undefined): boolean {
    return proto3.util.equals(Id, a, b);
  }
}